fixed gcc4.6 errors, added --no-auto-append

+++ src/util/strict_fstream.hpp
@@ -58,7 +58,7 @@ struct static_method_holder
                 "binary"
             };
         std::string res;
-        for (size_t i = 0; i < n_modes; ++i)
+        for (size_t i = 0; i < (size_t)n_modes; ++i)
         {
             if (mode & mode_val_v[i])
             {
